Current progress

+++ src/main/kotlin/model/Point.kt
@@ -2,4 +2,4 @@ package model
 
 import java.math.BigInteger
 
-data class Point(val x: BigInteger, val y: BigInteger)
+data class Point(val x: Int, val y: BigInteger)
